requirements.txt
absl-py==2.1.0
altair==5.2.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
attrs==23.2.0
autocfg==0.0.8
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
decorator==4.4.2
defusedxml==0.7.1
dm-tree==0.1.8
efficientnet==1.0.0
entrypoints==0.4
Faker==22.6.0
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
gluoncv==0.10.5.post0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.17.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-colab==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
grpcio==1.60.1
h5py==3.10.0
htbuilder==0.6.2
httplib2==0.22.0
idna==3.6
image-classifiers==1.0.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
ipykernel==5.5.6
ipython==7.34.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab_pygments==0.3.0
keras==2.15.0
Keras-Applications==1.0.8
kiwisolver==1.4.5
kornia==0.7.1
lapx==0.5.5
lazy_loader==0.3
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
mxnet==1.8.0
names==0.3.0
namex==0.0.7
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
portpicker==1.5.2
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.4
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
segmentation-models==1.0.1
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.31.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
supervision==0.18.0
svgpath2mpl==1.0.0
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
tf-keras==2.15.0
thop==0.1.1.post2209072238
tifffile==2024.1.30
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.9
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
yacs==0.1.8
zipp==3.17.0
